+++ pymodbus/pdu/decoders.py
@@ -2,7 +2,6 @@
 from __future__ import annotations
 
 import pymodbus.pdu.bit_read_message as bit_r_msg
-import pymodbus.pdu.bit_write_message as bit_w_msg
 import pymodbus.pdu.diag_message as diag_msg
 import pymodbus.pdu.file_message as file_msg
 import pymodbus.pdu.mei_message as mei_msg
@@ -22,10 +21,10 @@ class DecodePDU:
         (bit_r_msg.ReadDiscreteInputsRequest, bit_r_msg.ReadDiscreteInputsResponse),
         (reg_r_msg.ReadInputRegistersRequest, reg_r_msg.ReadInputRegistersResponse),
         (bit_r_msg.ReadCoilsRequest, bit_r_msg.ReadCoilsResponse),
-        (bit_w_msg.WriteMultipleCoilsRequest, bit_w_msg.WriteMultipleCoilsResponse),
+        (bit_r_msg.WriteMultipleCoilsRequest, bit_r_msg.WriteMultipleCoilsResponse),
         (reg_w_msg.WriteMultipleRegistersRequest, reg_w_msg.WriteMultipleRegistersResponse),
         (reg_w_msg.WriteSingleRegisterRequest, reg_w_msg.WriteSingleRegisterResponse),
-        (bit_w_msg.WriteSingleCoilRequest, bit_w_msg.WriteSingleCoilResponse),
+        (bit_r_msg.WriteSingleCoilRequest, bit_r_msg.WriteSingleCoilResponse),
         (reg_r_msg.ReadWriteMultipleRegistersRequest, reg_r_msg.ReadWriteMultipleRegistersResponse),
         (diag_msg.DiagnosticStatusRequest, diag_msg.DiagnosticStatusResponse),
         (o_msg.ReadExceptionStatusRequest, o_msg.ReadExceptionStatusResponse),
