+++ test/sub_current/test_remote_datastore.py
@@ -7,8 +7,7 @@ import pytest
 from pymodbus.datastore.remote import RemoteSlaveContext
 from pymodbus.exceptions import NotImplementedException
 from pymodbus.pdu import ExceptionResponse
-from pymodbus.pdu.bit_read_message import ReadCoilsResponse
-from pymodbus.pdu.bit_write_message import WriteMultipleCoilsResponse
+from pymodbus.pdu.bit_read_message import ReadCoilsResponse, WriteMultipleCoilsResponse
 from pymodbus.pdu.register_read_message import ReadInputRegistersResponse
